fix: #10 refactor code
