+++ data/db.json
@@ -13,22 +13,7 @@
       "password": "1234abcd"
     },
     {
-      "id": "bb0b167a-9885-4bb2-bde8-0a8e4ce02248",
-      "username": "",
-      "password": ""
-    },
-    {
-      "id": "420ef1b1-39bc-4426-b3ae-09710439a46f",
-      "username": "",
-      "password": ""
-    },
-    {
-      "id": "62d6ee71-8fed-45c2-81b8-839ed110c81b",
-      "username": "",
-      "password": ""
-    },
-    {
-      "id": "bf847428-88eb-4c81-a4f5-6150dc2d6c61",
+      "id": "8bbb796f-aeff-4f89-85d7-ba296cc8e9f9",
       "username": "",
       "password": ""
     }
