+++ react-basic-practice/.eslintrc.cjs
@@ -20,7 +20,6 @@ module.exports = {
     'react/prop-types': 'off',
     'react/function-component-definition': 'off',
     'no-restricted-syntax': 'off',
-    'import/prefer-default-export': 'off',
     'no-unused-expressions': 'off',
   },
 };
