+++ react-basic-practice/src/helpers/validate.js
@@ -1,0 +1,9 @@
+// import REGEXP from '../../constants/regexp';
+
+const Validate = (dataInput) => {
+  Object.entries(dataInput).forEach(([key, value]) => {
+    console.log([key, value]);
+  });
+};
+
+export default Validate;
